+++ docs/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+[![Versione Italiana](https://img.shields.io/badge/Italiano-Versione-green?style=flat-square)](CHANGELOG_IT.md)
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),